+++ marrow/src/_with_latest_arrow/arrow.rs
@@ -1,4 +1,5 @@
 //! a module that mirrors the arrow package for the most recent release
+#![allow(unused)]
 
 // arrow-version: replace: use arrow_array_{version} as _arrow_array;
 use arrow_array_53 as _arrow_array;
@@ -8,12 +9,28 @@ use arrow_schema_53 as _arrow_schema;
 
 pub mod array {
     pub use super::_arrow_array::array::{
-        make_array, Array, ArrayRef, ArrowPrimitiveType, BooleanArray, DictionaryArray,
-        FixedSizeBinaryArray, FixedSizeListArray, GenericBinaryArray, GenericListArray,
-        GenericStringArray, Int32Array, LargeStringArray, MapArray, NullArray, OffsetSizeTrait,
-        PrimitiveArray, StringArray, StructArray, UnionArray,
+        make_array, Array, ArrayRef, ArrowPrimitiveType, DictionaryArray, GenericBinaryArray,
+        GenericListArray, GenericStringArray, OffsetSizeTrait, PrimitiveArray,
     };
     pub use super::_arrow_array::RecordBatch;
+
+    // specialized arrays
+    pub use super::_arrow_array::array::{
+        BinaryArray, BooleanArray, Date32Array, Date64Array, DurationMicrosecondArray,
+        DurationMillisecondArray, DurationNanosecondArray, DurationSecondArray,
+        FixedSizeBinaryArray, FixedSizeListArray, Float16Array, Float32Array, Float64Array,
+        Int16Array, Int32Array, Int64Array, Int8Array, LargeBinaryArray, LargeStringArray,
+        MapArray, NullArray, StringArray, StructArray, Time32MillisecondArray, Time32SecondArray,
+        Time64MicrosecondArray, Time64NanosecondArray, TimestampMicrosecondArray,
+        TimestampMillisecondArray, TimestampNanosecondArray, TimestampSecondArray, UInt16Array,
+        UInt32Array, UInt64Array, UInt8Array, UnionArray,
+    };
+
+    // specialized builders
+    pub use super::_arrow_array::builder::{
+        FixedSizeListBuilder, Int32Builder, LargeListBuilder, ListBuilder, MapBuilder,
+        StringBuilder,
+    };
 }
 pub mod datatypes {
     pub use super::_arrow_array::types::{