+++ marrow/src/_with_latest_arrow/mod.rs
@@ -1,0 +1,5 @@
+//! Internal do not use
+//!
+//! Functionality that is available when the latest arrow version is enabled.
+
+pub mod arrow;