+++ marrow/src/view.rs
@@ -3,12 +3,21 @@
 //! The views correspond 1:1 to the corresponding arrays.
 use half::f16;
 
-use crate::datatypes::{FieldMeta, TimeUnit};
+use crate::{
+    datatypes::{FieldMeta, MapMeta, TimeUnit},
+    error::{fail, ErrorKind, Result},
+};
+
+// assert that the `Array` implements the expected traits
+const _: () = {
+    trait AssertExpectedTraits: Clone + std::fmt::Debug + PartialEq + Send + Sync {}
+    impl<'a> AssertExpectedTraits for View<'a> {}
+};
 
 /// An array with borrowed data
 ///
 /// See [`Array`][crate::array::Array]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 #[non_exhaustive]
 pub enum View<'a> {
     /// See [`Array::Null`][crate::array::Array::Null]
@@ -72,7 +81,7 @@ pub enum View<'a> {
     /// See [`Array::Dictionary`][crate::array::Array::Dictionary]
     Dictionary(DictionaryView<'a>),
     /// See [`Array::Map`][crate::array::Array::Map]
-    Map(ListView<'a, i32>),
+    Map(MapView<'a>),
     /// See [`Array::DenseUnion`][crate::array::Array::DenseUnion]
     DenseUnion(DenseUnionView<'a>),
 }
@@ -80,7 +89,7 @@ pub enum View<'a> {
 /// A bitmap with an optional offset
 ///
 /// The `i`-th element is stored at bit `offset + i`.
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct BitsWithOffset<'a> {
     /// The offset of the bits
     pub offset: usize,
@@ -89,14 +98,14 @@ pub struct BitsWithOffset<'a> {
 }
 
 /// See [`NullArray`][crate::array::NullArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct NullView {
     /// See [`NullArray::len`][crate::array::NullArray::len]
     pub len: usize,
 }
 
 /// See [`BooleanArray`][crate::array::BooleanArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct BooleanView<'a> {
     /// See [`BooleanArray::len`][crate::array::BooleanArray::len]
     pub len: usize,
@@ -107,7 +116,7 @@ pub struct BooleanView<'a> {
 }
 
 /// See [`PrimitiveArray`][crate::array::PrimitiveArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct PrimitiveView<'a, T> {
     /// See [`PrimitiveArray::validity`][crate::array::PrimitiveArray::validity]
     pub validity: Option<BitsWithOffset<'a>>,
@@ -116,7 +125,7 @@ pub struct PrimitiveView<'a, T> {
 }
 
 /// See [`TimeArray`][crate::array::TimeArray]
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct TimeView<'a, T> {
     /// See [`TimeArray::unit`][crate::array::TimeArray::unit]
     pub unit: TimeUnit,
@@ -127,7 +136,7 @@ pub struct TimeView<'a, T> {
 }
 
 /// See [`TimestampArray`][crate::array::TimestampArray]
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct TimestampView<'a> {
     /// See [`TimestampArray::unit`][crate::array::TimestampArray::unit]
     pub unit: TimeUnit,
@@ -140,18 +149,65 @@ pub struct TimestampView<'a> {
 }
 
 /// See [`StructArray`][crate::array::StructArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct StructView<'a> {
     /// See [`StructArray::len`][crate::array::StructArray::len]
     pub len: usize,
     /// See [`StructArray::validity`][crate::array::StructArray::validity]
     pub validity: Option<BitsWithOffset<'a>>,
     /// See [`StructArray::fields`][crate::array::StructArray::fields]
-    pub fields: Vec<(View<'a>, FieldMeta)>,
+    pub fields: Vec<(FieldMeta, View<'a>)>,
+}
+
+/// See [`MapArray`][crate::array::MapArray]
+#[derive(Clone, Debug, PartialEq)]
+pub struct MapView<'a> {
+    /// See [`MapArray::validity`][crate::array::MapArray::validity]
+    pub validity: Option<BitsWithOffset<'a>>,
+    /// See [`MapArray::offsets`][crate::array::MapArray::offsets]
+    pub offsets: &'a [i32],
+    /// See [`MapArray::meta`][crate::array::MapArray::meta]
+    pub meta: MapMeta,
+    /// See [`MapArray::keys`][crate::array::MapArray::keys]
+    pub keys: Box<View<'a>>,
+    /// See [`MapArray::values`][crate::array::MapArray::values]
+    pub values: Box<View<'a>>,
+}
+
+impl<'a> MapView<'a> {
+    #[allow(unused)]
+    pub(crate) fn from_logical_view(
+        entries: View<'a>,
+        entries_name: String,
+        sorted: bool,
+        validity: Option<BitsWithOffset<'a>>,
+        offsets: &'a [i32],
+    ) -> Result<Self> {
+        let View::Struct(entries) = entries else {
+            fail!(ErrorKind::Unsupported, "Expected struct array");
+        };
+        let Ok(entries_fields) = <[(FieldMeta, View); 2]>::try_from(entries.fields) else {
+            fail!(ErrorKind::Unsupported, "Expected two entries");
+        };
+        let [(keys_meta, keys_view), (values_meta, values_view)] = entries_fields;
+
+        Ok(MapView {
+            validity,
+            offsets,
+            meta: MapMeta {
+                entries_name,
+                sorted,
+                keys: keys_meta,
+                values: values_meta,
+            },
+            keys: Box::new(keys_view),
+            values: Box::new(values_view),
+        })
+    }
 }
 
 /// See [`ListArray`][crate::array::ListArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct ListView<'a, O> {
     /// See [`ListArray::validity`][crate::array::ListArray::validity]
     pub validity: Option<BitsWithOffset<'a>>,
@@ -164,7 +220,7 @@ pub struct ListView<'a, O> {
 }
 
 /// See [`FixedSizeListArray`][crate::array::FixedSizeListArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct FixedSizeListView<'a> {
     /// See [`FixedSizeListArray::len`][crate::array::FixedSizeListArray::len]
     pub len: usize,
@@ -179,7 +235,7 @@ pub struct FixedSizeListView<'a> {
 }
 
 /// See [`BytesArray`][crate::array::BytesArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct BytesView<'a, O> {
     /// See [`BytesArray::validity`][crate::array::BytesArray::validity]
     pub validity: Option<BitsWithOffset<'a>>,
@@ -190,7 +246,7 @@ pub struct BytesView<'a, O> {
 }
 
 /// See [`FixedSizeBinaryArray`][crate::array::FixedSizeBinaryArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct FixedSizeBinaryView<'a> {
     /// See [`FixedSizeBinaryArray::n`][crate::array::FixedSizeBinaryArray::n]
     pub n: i32,
@@ -201,7 +257,7 @@ pub struct FixedSizeBinaryView<'a> {
 }
 
 /// See [`DecimalArray`][crate::array::DecimalArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct DecimalView<'a, T> {
     /// See [`DecimalArray::precision`][crate::array::DecimalArray::precision]
     pub precision: u8,
@@ -214,7 +270,7 @@ pub struct DecimalView<'a, T> {
 }
 
 /// See [`DictionaryArray`][crate::array::DictionaryArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct DictionaryView<'a> {
     /// See [`DictionaryArray::indices`][crate::array::DictionaryArray::indices]
     pub indices: Box<View<'a>>,
@@ -223,12 +279,12 @@ pub struct DictionaryView<'a> {
 }
 
 /// See [`DenseUnionArray`][crate::array::DenseUnionArray]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct DenseUnionView<'a> {
     /// See [`DenseUnionArray::types`][crate::array::DenseUnionArray::types]
     pub types: &'a [i8],
     /// See [`DenseUnionArray::offsets`][crate::array::DenseUnionArray::offsets]
     pub offsets: &'a [i32],
     /// See [`DenseUnionArray::fields`][crate::array::DenseUnionArray::fields]
-    pub fields: Vec<(i8, View<'a>, FieldMeta)>,
+    pub fields: Vec<(i8, FieldMeta, View<'a>)>,
 }